requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-coreapi==2.3.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-formtools==2.4
django-guardian==2.4.0
django-import-export==3.0.1
django-reversion==5.0.4
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
future==0.18.2
httplib2==0.21.0
idna==3.4
importlib-metadata==5.1.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.1
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.3.0
PyJWT==1.7.1
pyparsing==3.0.9
pytz==2022.6
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
zipp==3.11.0
